Fix DINO type mismatch by using float32 and remove obsolete sam3 script

+++ teacher_oracle.py
@@ -13,13 +13,16 @@ class TeacherOracle:
         
         # Grounding DINO for text-to-box
         self.dino_processor = AutoProcessor.from_pretrained(dino_id)
-        self.dino_model = AutoModelForZeroShotObjectDetection.from_pretrained(dino_id).to(self.device).to(torch.float16)
+        # Load in float32 to avoid bias/input type mismatch errors (float32 is safer for general use)
+        self.dino_model = AutoModelForZeroShotObjectDetection.from_pretrained(dino_id).to(self.device)
 
     def get_ground_truth(self, image, text_query, box_threshold=0.3, text_threshold=0.25):
         """
         Takes an image and a text query, returns Grounding DINO's predicted BBox.
         """
+        # Ensure inputs are in float32 to match model weights
         inputs = self.dino_processor(images=image, text=text_query, return_tensors="pt").to(self.device)
+        
         with torch.no_grad():
             outputs = self.dino_model(**inputs)
